Fix processing of NaN values when OutputParserException
The RAGAS framework returns NaN when it encounters malformed output from
the LLM. The malformed output is accompanied by an OutputParserException
in the logs, but this exception is caught internally.

The NaN causes later failure during the generation of statistics like
standard deviation at the end of the evaluation and ultimately causes
no results to be obtained from the evaluation when the malformed output
is encountered by RAGAS.

This commit fixes this issue by checking whether NaN was returned from
RAGAS and, if so, ensures that the evaluate() function returns None, as
in other cases of failure. This ensures that NaN does not reach the
computation of the final statistics.

Resolves: #44